Split ProjectNavigatorView out of SidebarViewController.swift
Move the Project Navigator section (FileNode, ProjectNavigatorView and its
outline/menu extensions, DirectoryWatcher) into its own file. The only code
change: QuietSelectionRowView loses `private` because both outline views
use it.
